+++ CHANGELOG.md
@@ -6,6 +6,11 @@ title: Changelog
 
 ## Unreleased (????-??-??)
 
+* `<cc-button>`: Add a circle form when in hide-text with an image
+* New component:
+  * `<cc-pricing-table>`
+  * `<cc-pricing-product>` (with smart definition for add-ons and runtimes)
+
 ### For devs
 
 * i18n: simplify formatter system and add JSDoc